Move test to test_exit_codes
