+++ test/mitmproxy/addons/test_termlog.py
@@ -1,11 +1,8 @@
-import builtins
 import io
-import pytest
 
 from mitmproxy import log
 from mitmproxy.addons import termlog
 from mitmproxy.test import taddons
-from mitmproxy.utils import exit_codes
 
 
 def test_output(capsys):
@@ -31,18 +28,3 @@ def test_styling(monkeypatch) -> None:
         t.add_log(log.LogEntry("hello world", "info"))
 
     assert f.getvalue() == "\x1b[22mhello world\x1b[0m\n"
-
-
-def test_error_exit(monkeypatch) -> None:
-    def _raise(*args, **kwargs):
-        raise OSError
-
-    monkeypatch.setattr(builtins, "print", _raise)
-
-    t = termlog.TermLog()
-    with taddons.context(t) as tctx:
-        tctx.configure(t)
-        with pytest.raises(SystemExit) as exc_info:
-            t.add_log(log.LogEntry("error", "error"))
-
-        assert exc_info.value.args[0] == exit_codes.CANNOT_PRINT
